requirements.txt
Django==4.0.6
django-extensions==3.2.0
ipython==8.4.0
django-import-export==2.8.0

django-crispy-forms==1.14.0
crispy-bootstrap5==0.6

beautifulsoup4==4.11.1

# dev
django-debug-toolbar==3.5.0
django-admin-rangefilter==0.8.6
django-filter==22.1
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.0
drf-yasg

# flake8
flake8==4.0.1
flake8-debugger==4.1.2
flake8-print==5.0.0
flake8-builtins==1.5.3

# pytest
pytest==7.1.3
pytest-django==4.5.2
pytest-cov==4.0.0
pytest-mock==3.9.0

# 3d
rabbitmq
celery~=5.2.7
requests~=2.28.1